Crate ranker

Ranker is library, responsible for generating monitor Is is used in both apiserver and invoker

Structs

Cell
Represents one cell in monitor

Config

Monitor

PartyId

PartyRow

PartyStats
Represents some properties of row, describing party

ProblemConfig

ProblemId

ProblemStats
Represents some statistics of problem

Run

RunId

StatsRow

SubtaskId

Enums

PenaltyAggregation

ProblemScoreAggregationTarget

RunScoreAggregation
Determines how total score for problem is calculated

RunScoreAggregationTarget
Determines which runs will be used to calculate total score for problem.

Functions

build_cell

build_monitor
Builds a Monitor, given list of all runs

build_party_stats

Type Definitions

Score